Refactor compiler related classes.
- languageserver.Compiler is now compiler.ScalacProvider
- remove unused parameters to ScalacProvider
- remove unused classes
