+++ metaserver/src/main/scala/scala/meta/languageserver/compiler/SignatureHelpProvider.scala
@@ -2,7 +2,6 @@ package scala.meta.languageserver.compiler
 
 import scala.annotation.tailrec
 import scala.meta.languageserver.compiler.CompilerUtils._
-import scala.meta.languageserver.Compiler
 import scala.reflect.internal.util.Position
 import scala.tools.nsc.interactive.Global
 import com.typesafe.scalalogging.LazyLogging
@@ -13,7 +12,7 @@ import langserver.types.SignatureInformation
 object SignatureHelpProvider extends LazyLogging {
   def empty: SignatureHelp = SignatureHelp(Nil, None, None)
   def signatureHelp(compiler: Global, cursor: Cursor): SignatureHelp = {
-    val unit = Compiler.addCompilationUnit(
+    val unit = ScalacProvider.addCompilationUnit(
       global = compiler,
       code = cursor.contents,
       filename = cursor.uri,
